// client/src/app/api/models/article.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FileSystem } from '../models/file-system';
export interface Article {

  /**
   * The folder structure found in the articles folder
   */
  files?: Array<FileSystem>;

  /**
   * Article number
   */
  number?: string;
}